# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for ISHC project.",
        "title": "ISHC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/volunteers": {
            "get": {
                "description": "Get details of all volunteers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volunteer"
                ],
                "summary": "Get all volunteers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VolunteerInfo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new volunteer with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volunteer"
                ],
                "summary": "Create a new volunteer",
                "parameters": [
                    {
                        "description": "Volunteer payload",
                        "name": "volunteer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VolunteerInfo"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/volunteers/{id}": {
            "get": {
                "description": "Get details of a volunteer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volunteer"
                ],
                "summary": "Get a volunteer by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Volunteer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VolunteerInfo"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a volunteer with the given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volunteer"
                ],
                "summary": "Update a volunteer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Volunteer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Volunteer payload",
                        "name": "volunteer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VolunteerInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VolunteerInfo"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a volunteer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volunteer"
                ],
                "summary": "Delete a volunteer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Volunteer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/volunteers/{id}/profile_photo": {
            "post": {
                "description": "Set profile photo for a volunteer by ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volunteer"
                ],
                "summary": "Set volunteer profile photo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Volunteer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Profile Photo URL",
                        "name": "profile_photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.CustomTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
...
        "models.VolunteerInfo": {
            "type": "object",
            "properties": {
                "birthday": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "checkin_date": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "checkout_date": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "client_id": {
                    "type": "integer"
                },
                "created": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "created_by": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_card": {
                    "type": "string"
                },
                "imgset_dir": {
                    "type": "string"
                },
                "is_active": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "org_id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "profile_photo": {
                    "type": "string"
                },
                "remove": {
                    "type": "string"
                },
                "updated": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        }
    }
}